+++ src/MailCarrierServiceProvider.php
@@ -6,7 +6,7 @@ use Filament\Support\Assets\Js;
 use Filament\Support\Facades\FilamentAsset;
 use Illuminate\Support\Facades\Event;
 use MailCarrier\Commands\InstallCommand;
-use MailCarrier\Commands\LogRetryCommand;
+use MailCarrier\Commands\RetryCommand;
 use MailCarrier\Commands\SocialCommand;
 use MailCarrier\Commands\TokenCommand;
 use MailCarrier\Commands\UpgradeCommand;
@@ -44,7 +44,7 @@ class MailCarrierServiceProvider extends PackageServiceProvider
                 SocialCommand::class,
                 UserCommand::class,
                 TokenCommand::class,
-                LogRetryCommand::class,
+                RetryCommand::class,
             ])
             ->hasMigrations([
                 '1_create_users_table',
